rename base class file name
