+++ ctitoolbox/src/feedback/file_management.py
@@ -3,7 +3,7 @@ import base64
 
 import ArbinCTI.Core as ArbinCTI # type: ignore
 
-from ctitoolbox.src.feedback.feedback_base import DictReprBase
+from ctitoolbox.src.feedback.dict_repr_base import DictReprBase
 
 """""""""""""""""""""""""""
 File Management Feedback
