+++ ctitoolbox/src/feedback/system.py
@@ -2,7 +2,7 @@ from enum import IntEnum
 
 import ArbinCTI.Core as ArbinCTI # type: ignore
 
-from ctitoolbox.src.feedback.feedback_base import DictReprBase
+from ctitoolbox.src.feedback.dict_repr_base import DictReprBase
 """""""""""""""""""""""""""
 System & Connection
 - GetSerailNumberFeedback
